refactor: extract EventDaySelector from add_event_sheet
Moves _DaySelector (private) to lib/components/calendar/widgets/event_day_selector.dart
as public EventDaySelector widget.
