+++ lib/components/calendar/add_event_sheet.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:marquer/components/calendar/widgets/event_day_selector.dart';
 import 'package:marquer/components/shared/color_picker_row.dart';
 import 'package:marquer/providers/calendar/calendar_day_events_provider.dart';
 import 'package:marquer/utils/format.dart';
@@ -142,7 +143,7 @@ class _AddEventSheetState extends ConsumerState<AddEventSheet> {
                 Text(widget.initialName != null ? 'Edit Event' : 'Add Event', style: Theme.of(context).textTheme.titleMedium),
                 const SizedBox(height: 12),
                 if (widget.weekDays != null) ...[
-                  _DaySelector(
+                  EventDaySelector(
                     days: widget.weekDays!,
                     selected: _selectedDate,
                     colorScheme: colorScheme,
@@ -203,69 +204,6 @@ class _AddEventSheetState extends ConsumerState<AddEventSheet> {
           ),
         ),
       ),
-    );
-  }
-}
-
-class _DaySelector extends StatelessWidget {
-  final List<DateTime> days;
-  final DateTime? selected;
-  final ColorScheme colorScheme;
-  final List<String> dayLabels;
-  final void Function(DateTime) onSelect;
-
-  const _DaySelector({
-    required this.days,
-    required this.selected,
-    required this.colorScheme,
-    required this.dayLabels,
-    required this.onSelect,
-  });
-
-  bool _isSameDay(DateTime a, DateTime b) =>
-      a.year == b.year && a.month == b.month && a.day == b.day;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: List.generate(days.length, (i) {
-        final day = days[i];
-        final isSelected = selected != null && _isSameDay(day, selected!);
-        return Expanded(
-          child: GestureDetector(
-            onTap: () => onSelect(day),
-            child: Column(
-              children: [
-                Text(
-                  dayLabels[i],
-                  style: TextStyle(
-                    fontSize: 11,
-                    color: colorScheme.onSurface.withValues(alpha: 0.5),
-                  ),
-                ),
-                const SizedBox(height: 2),
-                Container(
-                  width: 28,
-                  height: 28,
-                  decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    color: isSelected ? colorScheme.primary : Colors.transparent,
-                  ),
-                  alignment: Alignment.center,
-                  child: Text(
-                    '${day.day}',
-                    style: TextStyle(
-                      fontSize: 13,
-                      fontWeight: FontWeight.w600,
-                      color: isSelected ? colorScheme.onPrimary : colorScheme.onSurface,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      }),
     );
   }
 }
